windows_deploy/backend/systems/visual/dreamer.py
import logging
import asyncio
import httpx
import json
import uuid
import websocket
from pathlib import Path

logger = logging.getLogger("Anuu_Dreamer")

class Dreamer:
    """
    The Visual Cortex: Connects to ComfyUI for Image/Video Generation.
    """
    def __init__(self, comfy_host: str = "127.0.0.1", port: int = 8188):
        self.server_address = f"{comfy_host}:{port}"
        self.client_id = str(uuid.uuid4())
        self.ws = None

    def queue_prompt(self, workflow: dict):
        p = {"prompt": workflow, "client_id": self.client_id}
        data = json.dumps(p).encode('utf-8')
        import urllib.request
        req = urllib.request.Request(f"http://{self.server_address}/prompt", data=data)
        return json.loads(urllib.request.urlopen(req).read())

    async def generate(self, prompt: str, workflow_json_path: str = None) -> str:
        """
        Simplified generation wrapper. 
        In v0.4, we assume a basic workflow is available or we construct a simple one.
        """
        if not workflow_json_path:
             # Construct a very basic T2I workflow dynamically if no file provided
             # This is a placeholder for the complex JSON ComfyUI expects.
             # Ideally, we load a pre-saved "flux_dev_api.json"
             logger.warning("No workflow provided to Dreamer. returning abstract hallucination.")
             return "Error: No neural pathway (workflow) found for this vision."
        
        try:
             # Load workflow, inject prompt, send to api
             return "Vision queued..."
        except Exception as e:
             return f"Vision Failed: {e}"

dreamer = Dreamer()
