windows_deploy/requirements.txt
fastapi
uvicorn
websockets
watchfiles
rich
ddgs
python-multipart
httpx
requests
colorama
